requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.2.1
python-dotenv==1.0.0
motor==3.3.2
httpx==0.25.2
email-validator==2.1.1
bcrypt==4.0.1
numpy==1.26.4
chromadb==0.4.24
google-generativeai
aiohttp
sentence-transformers
model2vec
einops